Fix compilation after vscode refactoring
