+++ src/odo.ts
@@ -25,7 +25,8 @@ import { Project } from './odo/project';
 import { ComponentsJson, NotAvailable } from './odo/component';
 import { Url } from './odo/url';
 import { Service, ServiceOperatorShortInfo } from './odo/service';
-import { Command, CommandText } from './odo/command';
+import { CommandText } from './base/command';
+import { Command } from './odo/command';
 import { VsCommandError } from './vscommand';
 import { Storage } from './odo/storage';
 import bs = require('binary-search');
