+++ test/integration/odo.test.ts
@@ -11,7 +11,8 @@ import * as path from 'path';
 import { Uri, window, commands, extensions } from 'vscode';
 import * as odo from '../../src/odo';
 import { Cluster } from '../../src/openshift/cluster';
-import { Command, CommandText } from '../../src/odo/command';
+import { CommandText } from '../../src/base/command';
+import { Command } from '../../src/odo/command';
 import { Component } from '../../src/openshift/component';
 import { AddWorkspaceFolder } from '../../src/util/workspace';
 import { ComponentTypeAdapter } from '../../src/odo/componentType';
