+++ test/unit/openshift/project.test.ts
@@ -8,7 +8,8 @@ import * as chai from 'chai';
 import * as sinonChai from 'sinon-chai';
 import * as sinon from 'sinon';
 import { OdoImpl, ContextType } from '../../../src/odo';
-import { Command, CommandText } from '../../../src/odo/command';
+import { CommandText } from '../../../src/base/command';
+import { Command } from '../../../src/odo/command';
 import { TestItem } from './testOSItem';
 import { Project } from '../../../src/openshift/project';
 import OpenShiftItem from '../../../src/openshift/openshiftItem';
